Adds executors to command interpreter
An executor is an object on which we can invoke commands by mapping a
command line to an object method using reflection. This allows to use
any object as a command executor for a shell application.

Adds executor registration.
Adds describe method to provide help for calling an executor method.
Updates invoke to match method based on name and parameter count.
Updates shell help command to use executor description.